# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "generate": {
            "default": false,
            "title": "Generate",
            "type": "boolean"
          },
          "k": {
            "default": 4,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "K",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "query": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AtlasChatRequest": {
        "properties": {
          "columns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "declared_causal_sufficiency": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Declared Causal Sufficiency"
          },
          "max_retries": {
            "default": 1,
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Max Retries",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "required": [
          "dataset_id"
        ],
        "title": "AtlasChatRequest",
        "type": "object"
      },
      "AtlasExperimentRequest": {
        "properties": {
          "columns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "declared_causal_sufficiency": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Declared Causal Sufficiency"
          },
          "max_lag": {
            "default": 1,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Lag",
            "type": "integer"
          },
          "max_methods": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 2.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 3,
            "title": "Max Methods"
          },
          "max_rows": {
            "anyOf": [
              {
                "minimum": 50.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Rows"
          },
          "methods": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Methods"
          },
          "n_bootstrap": {
            "default": 3,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "N Bootstrap",
            "type": "integer"
          },
          "use_assumption_soft_filter": {
            "default": true,
            "title": "Use Assumption Soft Filter",
            "type": "boolean"
          }
        },
        "required": [
          "dataset_id"
        ],
        "title": "AtlasExperimentRequest",
        "type": "object"
      },
      "BenchmarkRequest": {
        "properties": {
          "index_change": {
            "default": 250,
            "minimum": 0.0,
            "title": "Index Change",
            "type": "integer"
          },
          "max_lag": {
            "default": 2,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Lag",
            "type": "integer"
          },
          "methods": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Methods"
          },
          "n_bootstrap": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "N Bootstrap",
            "type": "integer"
          },
          "n_samples": {
            "default": 500,
            "maximum": 5000.0,
            "minimum": 100.0,
            "title": "N Samples",
            "type": "integer"
          },
          "noise_multiplier": {
            "default": 3.0,
            "minimum": 0.0,
            "title": "Noise Multiplier",
            "type": "number"
          }
        },
        "title": "BenchmarkRequest",
        "type": "object"
      },
...
      "CreateRunRequest": {
        "properties": {
          "columns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "decomposition_period": {
            "anyOf": [
              {
                "minimum": 2.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decomposition Period"
          },
          "ensemble_threshold": {
            "default": 0.5,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Ensemble Threshold",
            "type": "number"
          },
          "expert_knowledge": {
            "items": {
              "$ref": "#/components/schemas/ExpertRule"
            },
            "title": "Expert Knowledge",
            "type": "array"
          },
          "make_stationary": {
            "default": true,
            "title": "Make Stationary",
            "type": "boolean"
          },
          "max_lag": {
            "default": 2,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Lag",
            "type": "integer"
          },
          "methods": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Methods"
          },
          "n_bootstrap": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Bootstrap"
          },
          "normalize": {
            "default": true,
            "title": "Normalize",
            "type": "boolean"
          },
          "objective": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "panel_evidence": {
            "default": true,
            "title": "Panel Evidence",
            "type": "boolean"
          },
          "parallel_jobs": {
            "anyOf": [
              {
                "maximum": 8.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parallel Jobs"
          },
          "quick_mode": {
            "default": false,
            "title": "Quick Mode",
            "type": "boolean"
          },
          "selected_relations": {
            "anyOf": [
              {
                "items": {
                  "maxItems": 2,
                  "minItems": 2,
                  "prefixItems": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "string"
                    }
                  ],
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Relations"
          },
          "trajectory_index": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trajectory Index"
          }
        },
        "required": [
          "dataset_id"
        ],
        "title": "CreateRunRequest",
        "type": "object"
      },
...
      "ExpertRule": {
        "properties": {
          "confidence": {
            "default": 0.5,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "constraint": {
            "default": "soft",
            "enum": [
              "soft",
              "hard"
            ],
            "title": "Constraint",
            "type": "string"
          },
          "lag": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lag"
          },
          "prior_probability": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prior Probability"
          },
          "relation": {
            "default": "weak",
            "enum": [
              "strong",
              "weak",
              "inverse",
              "none"
            ],
            "title": "Relation",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "source",
          "target"
        ],
        "title": "ExpertRule",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReplicatedValidationRequest": {
        "properties": {
          "columns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "decomposition_period": {
            "anyOf": [
              {
                "minimum": 2.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decomposition Period"
          },
          "ensemble_threshold": {
            "default": 0.5,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Ensemble Threshold",
            "type": "number"
          },
          "max_lag": {
            "default": 2,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Lag",
            "type": "integer"
          },
          "methods": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Methods"
          },
          "min_confirmatory_replicates": {
            "default": 10,
            "minimum": 2.0,
            "title": "Min Confirmatory Replicates",
            "type": "integer"
          },
          "minimum_precision_gain": {
            "default": 0.05,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Minimum Precision Gain",
            "type": "number"
          },
          "minimum_win_rate": {
            "default": 0.7,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Minimum Win Rate",
            "type": "number"
          },
          "n_bootstrap": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Bootstrap"
          },
          "n_replicates": {
            "default": 10,
            "maximum": 100.0,
            "minimum": 2.0,
            "title": "N Replicates",
            "type": "integer"
          },
          "quick_mode": {
            "default": false,
            "title": "Quick Mode",
            "type": "boolean"
          },
          "replicate_seed": {
            "default": 2029,
            "title": "Replicate Seed",
            "type": "integer"
          },
          "significance_level": {
            "default": 0.05,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Significance Level",
            "type": "number"
          },
          "statistical_bootstraps": {
            "default": 10000,
            "maximum": 100000.0,
            "minimum": 100.0,
            "title": "Statistical Bootstraps",
            "type": "integer"
          }
        },
        "required": [
          "dataset_id"
        ],
        "title": "ReplicatedValidationRequest",
        "type": "object"
      },
      "RunView": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "pipeline",
              "benchmark",
              "replicated_validation",
              "atlas_experiment",
              "atlas_chat"
            ],
            "title": "Kind",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "progress": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "params",
          "created_at",
          "started_at",
          "finished_at",
          "error",
          "progress"
        ],
        "title": "RunView",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Causal Discovery TS API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/atlas/algorithms/{algorithm_id}": {
      "get": {
        "operationId": "get_algorithm_api_v1_atlas_algorithms__algorithm_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "algorithm_id",
            "required": true,
            "schema": {
              "title": "Algorithm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Algorithm Api V1 Atlas Algorithms  Algorithm Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Algorithm",
        "tags": [
          "atlas"
        ]
      }
    },
    "/api/v1/atlas/ask": {
      "post": {
        "description": "RAG local: trechos recuperados por TF-IDF; resposta do Ollama apenas com generate=true.",
        "operationId": "ask_api_v1_atlas_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ask Api V1 Atlas Ask Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "atlas"
        ]
      }
    },
...
    "/api/v1/runs": {
      "get": {
        "operationId": "list_runs_api_v1_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunView"
                  },
                  "title": "Response List Runs Api V1 Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Pipeline robusto (ENSEMBLE_AUTO) sobre um dataset.",
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/atlas-chat": {
      "post": {
        "description": "Selecao de metodos via chat local (Ollama), comparada ao filtro estatistico.",
        "operationId": "create_atlas_chat_api_v1_runs_atlas_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtlasChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Atlas Chat",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/atlas-experiment": {
      "post": {
        "description": "Perfil -> metodos compativeis -> selecao cega -> avaliacao pos-hoc.",
        "operationId": "create_atlas_experiment_api_v1_runs_atlas_experiment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtlasExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Atlas Experiment",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/benchmark": {
      "post": {
        "description": "Benchmark sintetico com ground truth conhecido + robustez a mudanca no regime de ruido.",
        "operationId": "create_benchmark_api_v1_runs_benchmark_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Benchmark",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/replicated-validation": {
      "post": {
        "description": "Validacao estatistica pareada em replicas independentes (~5 min por replica).",
        "operationId": "create_replicated_validation_api_v1_runs_replicated_validation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplicatedValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Replicated Validation",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "delete": {
        "operationId": "delete_run_api_v1_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Run",
        "tags": [
          "runs"
        ]
      },
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/result": {
      "get": {
        "operationId": "get_result_api_v1_runs__run_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Result Api V1 Runs  Run Id  Result Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Result",
        "tags": [
          "runs"
        ]
      }
    }
  }
}